scripts/requirements.txt
# AI Employee - Silver Tier Requirements

# Core dependencies
watchdog>=4.0.0

# Gmail API (for Gmail Watcher)
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.0.0

# Browser automation (for LinkedIn Watcher)
playwright>=1.40.0

# Install all:
#   pip install -r requirements.txt
#
# Then install Playwright browsers:
#   playwright install chromium
